Fix move navigation queuing
This was previously starting a new coroutine every physics frame, which
caused some horrible slowdown on longer running maps.

Now start the periodic nav refresh loop once at the start of the action.

Also leave a hook for the upcoming move_away to more easily use for a
different destination.