Prepare for columns in the tilemap
Adding obstacles to the tilemap allows them to more easily influence
pathfinding.

Otherwise move_away would try to run directly into the column in the
test level, and get stuck.

There is more to do here, like removing the standalone sprite and its
collision entirely, and fixing up y-sorting, but this is a start.