Directly use Skills in SkillTreeState instead of StringNames.
API doesn't change, but this makes it so it doesn't require SkillManager
to be available on startup before it calls _ready().

This was making us have to generate the skill dictionaries in the next
frame using call_deferred() in SkillTreeState._init(), which meant
SkillTreeState wasn't usable on the same frame it's created (e.g.
right after loading it in main.gd).

We'll see if this causes any issues :)

Closes #15.