Removes StatusComponent on HealthComponent.
It wasn't used yet. The idea was that statuses that cause direct damage
(e.g. poison) could directly interact with health, but it sounds better
that either HealthComponent takes care of that (by monitoring Status)
or some extra component like StatusActuator does that, by depending on
StatusComponent to be aware of new statuses, get HitEffect from them,
and apply them to Health.

This removes two of the last 4 cycles.

CYCLES=less

Commit Details

* Cycles before: 4, after: 2