Add an explicit 'prepare' stage to actions.
It can last a fixed amount of time, be explicily finished by the action
or not exist at all.

This will allow to have a condition for 'is preparing' that could e.g.
be used to use interrupting actions in a smart way.

In addition it makes it easier to just delay action effects.

Work for #57.

Commit Details

* Cycles before: 0, after: 0