A bunch of work for relic.
Add relic library state, separately from relic library.

Split out state from level_provider too (directly into run_save_state as
it's just an int).

Add a relic window.

This is still WIP and there is pending work, but checking it in for now
as it works fine and it's getting big.

Pending work is mostly:
* Allow to choose character that gets relic
* Add relic to character and have it take effect
* Add more relics, etc

Work for #9.

Commit Details

* Cycles before: 0, after: 0